🚚 Rename: theme 파일 js 전환
